refactor(dashboard): extract logs/processes endpoints to logs.rs (#1266)
routes.rs: 2881 -> 2613
logs.rs: NEW 277

Moves logs, read_tail, read_journal_logs, processes handlers.
Updates activity.rs import to point at new logs::read_tail.
